Common Utils file structure
Broke common utils into .c/.h files to match standard file structure.
